+++ src/components/ChessGameOverDialog/ChessGameOverDialog.tsx
@@ -1,9 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import Button from '@material-ui/core/Button';
-import Dialog from '@material-ui/core/Dialog';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogTitle from '@material-ui/core/DialogTitle';
-import Typography from '@material-ui/core/Typography';
+import { Button, Dialog, DialogContent, DialogTitle, Typography } from '@material-ui/core';
 
 import { getFullPieceColor } from '@/helpers';
 import { ChessGameOver, ChessGameOverType, ChessPieceColor } from '@/types';