+++ src/components/ChessPlayer/ChessPlayer.tsx
@@ -1,6 +1,6 @@
 import prettyMilliseconds from 'pretty-ms';
 import React from 'react';
-import Box from '@material-ui/core/Box';
+import { Box } from '@material-ui/core';
 
 import { ChessPlayer as ChessPlayerType } from '@/types';
 import useStyles from './styles';