+++ src/pages/index.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import Container from '@material-ui/core/Container';
-import Grid from '@material-ui/core/Grid';
+import { Container, Grid } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
 
 import SelectGame from '@/containers/SelectGame';